+++ fastify-jwt.d.ts
@@ -3,6 +3,7 @@ import '@fastify/jwt';
 declare module '@fastify/jwt' {
   interface FastifyJWT {
     user: {
+      customerId?: string;
       payload: {
         email: string;
         uuid: string;
@@ -15,7 +16,6 @@ declare module '@fastify/jwt' {
         sharedWorkspace: boolean;
         networkCredentials: {
           user: string;
-          pass: string;
         };
       };
     };
